requirements.txt
discord.py
python-dotenv
langchain
langchain-community
mlx